Made an CS for the classes and added a ProductBacklogClass
