+++ Scrum-Shot/Web.config
@@ -8,14 +8,14 @@
   <system.web>
     <compilation debug="true" targetFramework="4.7.2"/>
     <httpRuntime targetFramework="4.7.2"/>
-    <authentication mode="Forms">
+    <!--<authentication mode="Forms">
       <forms/>
     </authentication>
     <authorization>
       <deny users="?"/>
       <allow users="*"/>
-      <!--<deny users="*"/>-->
-    </authorization>
+      --><!--<deny users="*"/>--><!--
+    </authorization>-->
   </system.web>
   <connectionStrings>
     <add name="AppDB" connectionString="Driver={MySQL ODBC 3.51 Driver};
