+++ Scrum-Shot/index.aspx.cs
@@ -22,101 +22,31 @@ namespace Scrum_Shot
         Response.Redirect("~/Login.aspx");
     }
     }
-    public abstract class User
-    {
-        public User(int _userID, int _persID, string _firstname, string _lastname, string _email)
-        {
-            this.UserID = _userID;
-            this.PersID = _persID;
-            this.Firstname = _firstname;
-            this.Lastname = _lastname;
-            this.Email = _email;
-        }
-
-
-        protected int UserID { get;}
-        protected int PersID { get;}
-        protected string Firstname { get;}
-        protected string Lastname { get;}
-        protected string Email { get;}
-        protected string Password { get;}
-        protected int Priority { get; set; }
-        protected List<Project> Projects { get; }
-        public void AddProject(Project newProject)
-        {
-            Projects.Add(newProject);
-        }
-
-    }
-    public class Admin : User
-    {
-        public Admin(int _userID, int _persID, string _firstname, string _lastname, string _email) : base(_userID,_persID,_firstname,_lastname,_email)
-        {
-            this.Priority = 0;
-        }
-        
-        public static void AddPersonToProject(Project project, User userToAdd)
-        {
-            project.Participants.Add(userToAdd);
-            userToAdd.AddProject(project);
-        }
-    }
-        
-    public class ScrumMaster : User
-    {
-        public ScrumMaster(int _userID, int _persID, string _firstname, string _lastname, string _email) : base(_userID, _persID, _firstname, _lastname, _email)
-        {
-            Priority = 1;
-        }
-    }
-    public class ProductOwner : User
-    {
-        public ProductOwner(int _userID, int _persID, string _firstname, string _lastname, string _email) : base(_userID, _persID, _firstname, _lastname, _email)
-        {
-            Priority = 2;
-        }
-    }
-    public class TeamMember : User
-    {
-        public TeamMember(int _userID, int _persID, string _firstname, string _lastname, string _email) : base(_userID, _persID, _firstname, _lastname, _email)
-        {
-            Priority = 3;
-        }
-    }
+    
     public class Project
     {
-        int projectID;
-        string projectName;
         public List<User> participants = new List<User>();
 
-        public override string ToString()
+        public Project(string _projectName)
         {
-            return this.projectName.ToString();
+            //ProjectID muss vergeben werden, da es ein primärschlüssel ist. Darf nicht vom User eingegeben werden!
+            this.ProjectName = _projectName;
         }
-        public int PropertyID { get; set; }
+
+        public int ProjectID { get; set; }
         public string ProjectName { get; set; }
         public List<User> Participants { get; }
+        
+        public override string ToString()
+        {
+            return this.ProjectName.ToString();
+        }
+        public static void CreateProject(string projectName, ProductOwner productOwner)
+        {
+            //ist derjenige der ein Projekt erstellt der SCRUM-Master oder der Project-Owner? Oder kann auch ein User ein Projekt erstellen und den Scrum Master zuweisen=?
+            Project newProject = new Project(projectName);
+            productOwner.AddProject(newProject);
+        }
     }
 
-    public class ProductBacklog
-    {
-        List<ProductBacklogPoint> backlogPoints;
-
-    }
-    public class ProductBacklogPoint
-    {
-        string description;
-        int storyPoints;
-        bool inProgress;
-        bool done;
-        User editedBy;
-
-
-
-        public string Description{ get; set; }
-        public int StoryPoints { get; set; }
-        public bool InProgress { get; set; }
-        public bool Done { get; set; }
-        public User EditedBy{ get; set; }
-    }
 }
